+++ App.tsx
@@ -1,27 +1,13 @@
-import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import { initializeApp } from "firebase/app";
-import OnboardingScreens from './components/OnboardingScreens/OnboardingScreens';
+import OnboardingScreens from './src/screens/Onboarding/OnboardingScreens';
 
 // Import the functions you need from the SDKs you need
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
-
-
 export default function App() {
   return (
     <OnboardingScreens/>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
