+++ src/screens/Login/recoverPassword.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { StyleSheet, Text, TextInput, View, Image, TouchableOpacity } from 'react-native';
-import { TURQUOISE } from "../common/colors";
-import FirebaseInteractor from "../firebase/firebaseInteractor";
+import { TURQUOISE } from "../../common/colors";
+import FirebaseInteractor from "../../firebase/firebaseInteractor";
 
 let interactor = new FirebaseInteractor()
 
@@ -14,7 +14,7 @@ export default function RecoverPasswordPage({ goToSignIn }: RecoveryPasswordProp
 
     return (
         <View style={styles.container}>
-            <Image source={require("../assets/flow-icon.png")} style={styles.mainImage} />
+            <Image source={require("../../assets/flow-icon.png")} style={styles.mainImage} />
             <TextInput placeholderTextColor="#4D4661" value={email} onChangeText={setEmail} style={styles.textInput} placeholder="email" />
             <TouchableOpacity onPress={() => {
                 interactor.resetPassword(email).catch(console.log).then(console.log)
